requirements.txt
asgiref==3.5.0
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==2.3.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
